requirements.txt
ultralytics
opencv-python
numpy
fastapi
tensorflow
scikit-learn
matplotlib
langchain[groq]
groq
langchain
langchain_community
python-dotenv
pathlib